--- a/core/include/oneseismic/decoder.hpp
+++ b/core/include/oneseismic/decoder.hpp
@@ -8,6 +8,7 @@
 
 #include <oneseismic/messages.hpp>
 
+
 namespace one {
 
 /*