--- a/core/include/oneseismic/geometry.hpp
+++ b/core/include/oneseismic/geometry.hpp
@@ -3,6 +3,7 @@
 
 #include <array>
 #include <cassert>
+#include <iostream>
 #include <string>
 #include <vector>