… .github/workflows/ci.yaml
… .github/workflows/wheels.yml
… .travis.yml
… core/CMakeLists.txt
… core/tests/messages.cpp
… core/tests/testsuite.cpp
… python/CMakeLists.txt
… python/oneseismic/decoder.cpp
… python/pyproject.toml
… python/requirements-dev.txt
… python/setup.cfg
… python/setup.py